Get more data from GCP about images
The original proof of concept functions only retrieved the name of the
image from GCP but no additional data was included. This patch updates
those methods by adding in pertinent information from Google's image
list object that could be useful in the locator.

Google returns an unusual object (with more objects under attributes)
that is difficult to parse, so it required an additional function to
normalize the data and return it similarly to how we handle AWS and
Azure.

Signed-off-by: Major Hayden <major@redhat.com>