Use pre-commit in CI with sane defaults
The pre-commit config was originally copied from poetry's repo, but they
had lots of adjustments to the defaults. This commit restores some
defaults from the upstream projects.

Also, change CI to actually run the same pre-commit hooks that run on
developer workstations so we can test the same way in both places.

Signed-off-by: Major Hayden <major@redhat.com>