transformer: Provide only the latest images
If there are multiple images of the same name, use only the latest one.
This is a workaround of an issue with the CID API, that generates files
for these images at the same path and looses data about images as a result.
Let's throw away the data anyway, but this time intentionally.

Resolves: #492