V2ListRegionByVersion: Try simplifying
Remove some loops and build out a DataEntry-like structure from the
start. Then convert it to data entries.

The complexity came down enough that we can remove the ruff ignore line.

Update the test to check all results and add an extra google entry for 9
so we can verify it comes back with a count of 2.

Signed-off-by: Major Hayden <major@redhat.com>